# backend-fastapi/.docs/chatbot.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NutriAgent Chatbot API",
    "description": "AI 챗봇 API — AWS Bedrock(Claude) 기반, SSE 스트리밍 지원, 페르소나 시스템",
    "version": "1.0.0"
  },
  "tags": [],
  "paths": {
    "/api/v1/chat/personas": {
      "get": {
        "summary": "페르소나 목록 조회",
        "deprecated": false,
        "description": "사용 가능한 AI 페르소나 목록을 반환합니다.\n\n### 페르소나란?\n챗봇의 대화 스타일을 결정하는 설정입니다. 세션 생성 시 선택합니다.\n\n### 테스트 시나리오:\n  1. 정상 요청 → 200 OK, 페르소나 배열 반환",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {}
                  }
                },
                "example": [
                  {
                    "name": "friendly_pt",
                    "display_name": "친근한 트레이너",
                    "description": "따뜻하고 격려하는 스타일의 퍼스널 트레이너"
                  },
                  {
                    "name": "drill_sergeant",
                    "display_name": "강훈련 코치",
                    "description": "엄격하고 도전적인 스타일의 코치"
                  }
                ]
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/chat/sessions": {
      "post": {
        "summary": "대화 세션 생성",
        "deprecated": false,
        "description": "새 AI 챗봇 대화 세션을 생성합니다.\n\n### 세션이란?\n대화 이력을 유지하는 단위입니다. 하나의 세션 안에서 대화가 이어지며, 새 주제는 새 세션을 생성합니다.\n\n### 테스트 시나리오:\n  1. 정상 요청 (유효한 guest_id + persona) → 200 OK\n  2. 존재하지 않는 persona → 400 Bad Request\n  3. guest_id 누락 → 422 Unprocessable Entity",
        "tags": [],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cookie",
            "in": "header",
            "description": "Spring 세션 쿠키 (guest_id 추출용)",
            "required": false,
            "example": "{{cookie}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "guest_id": "guest_c8812f49-54b7-4858-85e6-65f98b4a410e",
                "persona": "friendly_pt"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "id": 1,
                  "guest_id": "guest_c8812f49-54b7-4858-85e6-65f98b4a410e",
                  "persona": "friendly_pt",
                  "title": null,
                  "created_at": "2026-04-16T09:00:00Z",
                  "updated_at": "2026-04-16T09:00:00Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "존재하지 않는 페르소나",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "detail": "Unknown persona: unknown_persona"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "대화 세션 목록 조회",
        "deprecated": false,
        "description": "사용자의 대화 세션 목록을 최신순으로 반환합니다.\n\n### 테스트 시나리오:\n  1. 세션이 있는 guest_id → 200 OK, 세션 배열 반환\n  2. 세션이 없는 guest_id → 200 OK, 빈 배열 반환",
        "tags": [],
        "parameters": [
          {
            "name": "guest_id",
            "in": "query",
            "description": "게스트 ID",
            "required": true,
            "example": "guest_c8812f49-54b7-4858-85e6-65f98b4a410e",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {}
                  }
                },
                "example": [
                  {
                    "id": 2,
                    "guest_id": "guest_c8812f49-54b7-4858-85e6-65f98b4a410e",
                    "persona": "friendly_pt",
                    "title": "오늘 점심 식단 상담",
                    "created_at": "2026-04-16T09:00:00Z",
                    "updated_at": "2026-04-16T09:10:00Z"
                  },
                  {
                    "id": 1,
                    "guest_id": "guest_c8812f49-54b7-4858-85e6-65f98b4a410e",
                    "persona": "drill_sergeant",
                    "title": null,
                    "created_at": "2026-04-15T08:00:00Z",
                    "updated_at": "2026-04-15T08:05:00Z"
                  }
                ]
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/chat/sessions/{session_id}/messages": {
      "get": {
        "summary": "대화 이력 조회",
        "deprecated": false,
        "description": "특정 세션의 메시지 이력을 시간순으로 반환합니다.\n\n> tool_calls / tool_results는 제외하고 텍스트 메시지만 반환됩니다.\n\n### 테스트 시나리오:\n  1. 본인 세션 조회 → 200 OK\n  2. 타인 세션 조회 → 403 Forbidden\n  3. 존재하지 않는 session_id → 404 Not Found",
        "tags": [],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "세션 ID",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "X-Guest-Id",
            "in": "header",
            "description": "게스트 ID",
            "required": true,
            "example": "guest_c8812f49-54b7-4858-85e6-65f98b4a410e",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {}
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "role": "user",
                    "content": "오늘 식단 어떻게 짜면 좋을까요?",
                    "created_at": "2026-04-16T09:00:00Z"
                  },
                  {
                    "id": 2,
                    "role": "assistant",
                    "content": "안녕하세요! 오늘 목표 칼로리를 먼저 확인해볼게요 😊",
                    "created_at": "2026-04-16T09:00:05Z"
                  }
                ]
              }
            },
            "headers": {}
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "detail": "Access denied"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "세션 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "detail": "Session not found"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "메시지 전송 (일반 응답)",
        "deprecated": false,
        "description": "메시지를 전송하고 AI 응답을 한 번에 받습니다.\n\n> 스트리밍이 필요한 경우 `/messages/stream` 엔드포인트를 사용하세요.\n\n### 테스트 시나리오:\n  1. 정상 메시지 전송 → 200 OK, AI 응답 반환\n  2. 빈 메시지 전송 → 422 Unprocessable Entity\n  3. 타인 세션으로 전송 → 403 Forbidden\n  4. 존재하지 않는 session_id → 404 Not Found",
        "tags": [],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "세션 ID",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Guest-Id",
            "in": "header",
            "description": "게스트 ID",
            "required": true,
            "example": "guest_c8812f49-54b7-4858-85e6-65f98b4a410e",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "message": "오늘 점심 추천해줘"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "response": "오늘 점심으로는 닭가슴살 샐러드를 추천드려요! 단백질이 풍부하고 칼로리도 적당해서 목표 달성에 도움이 될 거예요 😊"
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "detail": "Access denied"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "세션 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "detail": "Session not found"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/chat/sessions/{session_id}/messages/stream": {
      "post": {
        "summary": "메시지 전송 (SSE 스트리밍)",
        "deprecated": false,
        "description": "메시지를 전송하고 AI 응답을 Server-Sent Events로 실시간 스트리밍합니다.\n\n### SSE 이벤트 포맷\n```\ndata: {\"type\": \"content\", \"text\": \"닭가슴\"}\ndata: {\"type\": \"content\", \"text\": \"살은...\"}\ndata: {\"type\": \"done\"}\n```\n\n| 필드 | 설명 |\n|------|------|\n| `type: content` | 텍스트 청크 (실시간 출력) |\n| `type: done` | 스트림 종료 |\n\n### 프론트엔드 연동 예시\n```js\nconst res = await fetch(url, {\n  method: 'POST',\n  headers: { 'X-Guest-Id': guestId, 'Content-Type': 'application/json' },\n  body: JSON.stringify({ message: '오늘 점심 추천해줘' })\n});\nconst reader = res.body.getReader();\nconst decoder = new TextDecoder();\nwhile (true) {\n  const { done, value } = await reader.read();\n  if (done) break;\n  const text = decoder.decode(value);\n  // 'data: {...}' 파싱 후 UI 업데이트\n}\n```\n\n### 테스트 시나리오:\n  1. 정상 메시지 전송 → 200, SSE 스트림 수신\n  2. 타인 세션으로 전송 → 403 Forbidden\n  3. 존재하지 않는 session_id → 404 Not Found",
        "tags": [],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "세션 ID",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Guest-Id",
            "in": "header",
            "description": "게스트 ID",
            "required": true,
            "example": "guest_c8812f49-54b7-4858-85e6-65f98b4a410e",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "message": "오늘 점심 추천해줘"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE 스트림 응답 (Content-Type: text/event-stream)",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "data: {\"type\": \"content\", \"text\": \"오늘 점심으로는 \"}\n\ndata: {\"type\": \"content\", \"text\": \"닭가슴살 샐러드를 추천드려요!\"}\n\ndata: {\"type\": \"done\"}\n\n"
              }
            },
            "headers": {}
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "detail": "Access denied"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "세션 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "detail": "Session not found"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {}
  }
}